limo seems to work on cygwin now, wow

+++ Makefile
@@ -35,19 +35,27 @@ limo: $(OBJ) libs info
 limo-almost-static: $(OBJ) libs info
 	$(CC) $(OBJ) $(PROFILING) `./inline-cfg.sh` -lm -l:libgc.a -lpthread -l:libgmp.a -ldl -l:libreadline.a -l:libtermcap.a -rdynamic -o limo
 
+limo-wsl: limo-almost-static info
+
+libs:
+	make -C libs
+
+
+############################
+## cygwin
+
 limo-cygwin-exe: $(OBJ)
 	$(CC) $(BASEOBJ) $(PROFILING) -shared `./inline-cfg.sh` -lm -lgc -lpthread -lgmp -ldl -lreadline -rdynamic -o limodll.dll
 	$(CC) main.o $(PROFILING) limodll.dll -lm -lgc -lpthread -lgmp -ldl -lreadline -rdynamic -o limo.exe
 
 limo-cygwin: limo-cygwin-exe libs-cygwin info
 
-limo-wsl: limo-almost-static info
-
-libs:
-	make -C libs
-
 libs-cygwin:
-	make -C libs lib-cygwin
+	make -C libs libs-cygwin
+
+############################
+
+
 
 TAGS:
 	etags.emacs --regex='{c}/BUILTIN\(FUN\)?(\([^)]+\)/\2/' *.c *.h libs/*/*.c libs/*/*.h
@@ -71,12 +79,15 @@ realclean: clean
 install:
 	mkdir -p $(LIMO_PREFIX)
 	cp -f *.limo $(LIMO_PREFIX)
-#	cp -f libs/*/*.so $(LIMO_PREFIX)
+	cp -f libs/*/*.so $(LIMO_PREFIX) || true
+	cp -f libs/*/*.dll $(LIMO_PREFIX) || true
 	cp -f libs/*/*.ttf $(LIMO_PREFIX)
 	cp -rf limo-code/* $(LIMO_PREFIX)
 	cp -f docs/limo-info.info $(LIMO_PREFIX)
 	mkdir -p $(INSTALL_PREFIX)/bin
-	cp -f limo $(INSTALL_PREFIX)/bin
+	[ -e limo ] && cp -f limo $(INSTALL_PREFIX)/bin
+	[ -e limodll.dll ] && mv $(LIMO_PREFIX)/limodll.dll $(INSTALL_PREFIX)/bin   # yeah, I'm sorry I move this around so much
+	[ -e limo.exe ] && cp -f limo.exe $(INSTALL_PREFIX)/bin
 
 uninstall:
 	rm -rf $(LIMO_PREFIX)
